web/map_docker/requirements.txt
fastapi==0.115.12
pymongo==4.13.0
python-socketio==5.13.0
uvicorn==0.34.2
